Add new tokens to the annotations file.  Convert them appropriately.  Minor fix to page layout.

+++ doconversion.pl
@@ -18,29 +18,48 @@ while(my $line = <$file>){
     my ($id, $name, $count, $mask) = split /,/, $line;
 
     # convert file
-    my $infile = "$src/marker-$id.$src";
-    my $outfile = "$dest/marker-$id.$dest";
-    my $thumbfile = "thumb/marker-$id.$dest";
-    system("convert -density 150 -units PixelsPerInch $infile $outfile");
-
-    # apply mask
-    if($mask){
-        print "Masking $id based on $mask\n";
-        my $maskfile = "$src/marker-$mask.$src";
-        my $tmpfile = "$dest/orig-marker-$id.$dest";
-
-        system("mv $outfile $tmpfile");
-        system("convert \\( $maskfile -negate \\) $tmpfile $maskfile -composite $outfile");
-        unlink($tmpfile);
+    my $infile;
+    if($id =~ m/^\d\d\d$/){
+        $infile = "$src/marker-$id.$src";
+    }else{
+        $infile = "$src/$id.png";
+        $id =~ s/ /_/g;
     }
 
-    # trim whitespace
-    system("mogrify -fuzz 5% -trim +repage $outfile");
+    # Skip files that are not in this directory
+    if(-f $infile){
+        my $outfile = "$dest/marker-$id.$dest";
+        my $thumbfile = "thumb/marker-$id.$dest";
 
-    # generate thumbnail
-    my ($xres, $yres) = imgsize($outfile);
-    my $width = 30;
-    my $height = $yres * $width / $xres;
+        my $resize = "";
+        my ($xres, $yres) = imgsize($infile);
+        if($xres > 350){
+            my $width = 350;
+            my $height = $yres * $width / $xres;
+            $resize = "-resize ".$width."x".$height;
+        }
 
-    system("convert $outfile -resize $width"."x"."$height $thumbfile");
+        system("convert -density 150 -units PixelsPerInch -background white -flatten $resize '$infile' $outfile");
+
+        # apply mask
+        if($mask){
+            print "Masking $id based on $mask\n";
+            my $maskfile = "$src/marker-$mask.$src";
+            my $tmpfile = "$dest/orig-marker-$id.$dest";
+
+            system("mv $outfile $tmpfile");
+            system("convert \\( $maskfile -negate \\) $tmpfile $maskfile -composite $outfile");
+            unlink($tmpfile);
+        }
+
+        # trim whitespace
+        system("mogrify -fuzz 5% -trim +repage $outfile");
+
+        # generate thumbnail
+        ($xres, $yres) = imgsize($outfile);
+        my $width = 30;
+        my $height = $yres * $width / $xres;
+
+        system("convert $outfile -resize $width"."x"."$height $thumbfile");
+    }
 }
